created reusable component for proposal details
