+++ instances/widgets.treasury-factory.near/widget/pages/payments/Table.jsx
@@ -1,7 +1,9 @@
 const { href } = VM.require("${REPL_DEVHUB}/widget/core.lib.url") || {
   href: () => {},
 };
+
 const {
+  decodeBase64,
   getNearBalances,
   decodeProposalDescription,
   formatSubmissionTimeStamp,
@@ -161,18 +163,6 @@ const daoFTTokens = fetch(
 const nearBalances = getNearBalances(treasuryDaoID);
 
 const proposalPeriod = policy.proposal_period;
-
-function decodeBase64(encodedArgs) {
-  if (!encodedArgs) return null;
-  try {
-    const jsonString = Buffer.from(encodedArgs, "base64").toString("utf8");
-    const parsedArgs = JSON.parse(jsonString);
-    return parsedArgs;
-  } catch (error) {
-    console.error("Failed to decode or parse encodedArgs:", error);
-    return null;
-  }
-}
 
 const hasOneDeleteIcon =
   isPendingRequests &&
